[Feature] LiveRoom 및 User 엔티티 수정

Closes #1

See merge request s10-webmobile1-sub2/S10P12C203!2

+++ back/src/main/java/com/ssafy/sub/pjt/controller/UserController.java
@@ -23,7 +23,7 @@ public class UserController {
 
     @GetMapping("/{userId}")
     public ResponseEntity<?> getMyInfo(
-            @PathVariable final Long userId, @RequestBody final MyPageRequest myPageRequest) {
+            @PathVariable final Integer userId, @RequestBody final MyPageRequest myPageRequest) {
         userService.updateMyPageInfo(userId, myPageRequest);
         return ResponseEntity.noContent().build();
     }
